chore(genesis): De-dup Revm SpecId Helper (#142)
### Description

Introduces a `revm` feature flag to the `maili-genesis` crate.
Adds a helper method to the `RollupConfig` behind the `revm`
feature flag that exposes a `spec_id(timestamp)` method to get
the `revm::primitives::SpecId` for the given timestamp.

This is duplicated in a few places, and upstreams the functionality
into `maili`, specific for Optimism.

Closes #140