feat(genesis): Add optional interop message expiry window field (#208)
## Overview

Adds an optional `interop_message_expiry_window` field to the
`RollupConfig` type. As it stands, there is no information from the
`superchain-registry` to populate this field on a per-chain basis.
Similar to the channel timeout parameters, though, it can be overridden
from the default constant (1 hour to start, likely to be bumped post
EIP-2935.)